pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "fastapi-backend"
version = "0.1.0"
description = "Production-grade FastAPI backend with standardized structure"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn>=0.22.0",
    "sqlalchemy>=2.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "alembic>=1.11.0",
    "httpx>=0.24.0",
    "stamina>=22.1.0",
    "prometheus-client>=0.17.0",
    "python-dotenv>=1.0.0",
    "python-jose>=3.3.0",
    "passlib>=1.7.4",
    "email-validator>=2.0.0",
    "python-multipart>=0.0.6",  # Required for form data handling
    "supabase>=2.3.0",          # Supabase Python client
    "gotrue>=1.1.0",            # Supabase Auth client
    # Database drivers - SQLite is included by default
    # "psycopg2-binary>=2.9.5",  # Uncomment for PostgreSQL
]

[project.optional-dependencies]
dev = [
    "pytest>=7.3.1",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.10.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.3.0",
    "isort>=5.12.0",
    "mypy>=1.3.0",
    "pylint>=2.17.4",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
